requirements.txt
python-dotenv
openai
duckduckgo_search
pandas